crud methods completed

+++ src/main/java/gadgetStore/service/serviceImpl/UserServiceImpl.java
@@ -7,6 +7,8 @@ import gadgetStore.dto.userDto.response.UserResponse;
 import gadgetStore.entities.Favorite;
 import gadgetStore.entities.Product;
 import gadgetStore.entities.User;
+import gadgetStore.enums.Role;
+import gadgetStore.exceptions.ImpossibleException;
 import gadgetStore.repository.BasketRepo;
 import gadgetStore.repository.FavoriteRepo;
 import gadgetStore.repository.ProductRepo;
@@ -18,15 +20,12 @@ import org.springframework.stereotype.Service;
 
 import java.time.ZonedDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     private final UserRepo userRepo;
-    private final ProductRepo productRepo;
-    private final FavoriteRepo favoriteRepo;
-    private final JwtService jwtService;
-    private final BasketRepo basketRepo;
 
     @Override
     public UserResponse getById(Long id) {
@@ -41,7 +40,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public List<UserResponse> getAll() {
-         return userRepo.getAll();
+        return userRepo.getAll();
     }
 
     @Override
@@ -55,48 +54,23 @@ public class UserServiceImpl implements UserService {
 
         return SimpleResponse.builder()
                 .httpStatus(HttpStatus.OK)
-                .message("User with id '"+id+"' successfully updated")
+                .message("User with id '" + id + "' successfully updated")
                 .build();
     }
 
     @Override
     public SimpleResponse delete(Long id) {
         User user = userRepo.getUserByIdOrException(id);
+        if (user.getRole().equals(Role.ADMIN)){
+            throw new ImpossibleException("Impossible delete the admin!");
+        }
         userRepo.delete(user);
         return SimpleResponse.builder()
                 .httpStatus(HttpStatus.OK)
-                .message("User with id '"+ id +"' successfully deleted")
+                .message("User with id '" + id + "' successfully deleted")
                 .build();
     }
 
-    @Override
-    public SimpleResponse addProductToFavourite(Long productId) {
-        User user = jwtService.getAuthentication();
-        Product product = productRepo.getProductByIdOrException(productId);
-
-        Favorite favorite = Favorite.builder()
-                .product(product)
-                .user(user)
-                .build();
 
 
-        favoriteRepo.save(favorite);
-        user.getFavorites().add(favorite);
-        return SimpleResponse.builder()
-                .httpStatus(HttpStatus.OK)
-                .message("Product with id '" + productId + "' successfully added to favourite")
-                .build();
-    }
-
-    @Override
-    public SimpleResponse addProductToBasket(Long productId) {
-        User user = jwtService.getAuthentication();
-        user.getBasket().getProducts().add(productRepo.getProductByIdOrException(productId));
-        userRepo.save(user);
-
-        return SimpleResponse.builder()
-                .httpStatus(HttpStatus.OK)
-                .message("Product with id '" + productId + "' successfully added to basket")
-                .build();
-    }
 }
